+++ lib/Braintree/PaymentInstrumentType.php
@@ -14,5 +14,6 @@ class PaymentInstrumentType
     const VENMO_ACCOUNT       = 'venmo_account';
     const US_BANK_ACCOUNT     = 'us_bank_account';
     const IDEAL_PAYMENT       = 'ideal_payment';
+    const SAMSUNG_PAY_CARD    = 'samsung_pay_card';
 }
 class_alias('Braintree\PaymentInstrumentType', 'Braintree_PaymentInstrumentType');