+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.35.0
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
 * Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.
 * Fix issue where result objects could not be printed with `echo` (thanks @cmeon)